- Correcao da performance do app. Estava criando objetos desnecessarios para trabalhar com banco de dados. Limpei os helpers e as daos sempre que perderem referencia.

+++ app/src/main/java/br/org/unesco/appesca/bo/QuestaoBO.java
@@ -16,37 +16,39 @@ import br.org.unesco.appesca.model.Resposta;
  */
 public class QuestaoBO {
 
-    public void excluirQuestao(Questao questao, Context context){
+    QuestaoDAO questaoDAO;
+    RespostaDAO respostaDAO;
+    PerguntaDAO perguntaDAO;
+
+    public QuestaoBO(Context context){
+        questaoDAO = new QuestaoDAO(context);
+        respostaDAO = new RespostaDAO(context);
+        perguntaDAO = new PerguntaDAO(context);
+    }
+
+    public void excluirQuestao(Questao questao){
 
         if (questao != null) {
-            QuestaoDAO questaoDAO = new QuestaoDAO(context);
-            RespostaDAO respostaDAO = new RespostaDAO(context);
-
-            PerguntaDAO perguntaDAO = new PerguntaDAO(context);
-
             List<Pergunta> listaPerguntas = questao.getPerguntas();
 
             for (Pergunta pergunta : listaPerguntas) {
                 List<Resposta> listaRespostas = pergunta.getRespostas();
 
                 for (Resposta resposta : listaRespostas) {
-                    respostaDAO.delete(resposta.getId());
+                    if(resposta!=null && resposta.getId()!=null) {
+                        respostaDAO.delete(resposta.getId());
+                    }
                 }
                 perguntaDAO.deletePerguntaById(pergunta.getId());
             }
             questaoDAO.deleteQuestaoById(questao.getId());
-
         }
     }
 
-    public boolean temAlgumaResposta(Questao questao, Context context){
-
+    public boolean temAlgumaResposta(Questao questao){
         boolean ret = false;
 
         if(questao!=null){
-            RespostaDAO respostaDAO = new RespostaDAO(context);
-            PerguntaDAO perguntaDAO = new PerguntaDAO(context);
-
             List<Pergunta> listaPerguntas = perguntaDAO.findPerguntasByQuestao(questao.getId());
 
             for (Pergunta pergunta : listaPerguntas) {
@@ -57,7 +59,6 @@ public class QuestaoBO {
                 }
             }
         }
-
         return ret;
     }
 }
